+++ apps/manager/server/src/app.module.ts
@@ -5,9 +5,11 @@ import { IterationService } from "./iteration.service";
 import { RequirementController } from "./requirement.controller";
 import { RequirementService } from "./requirement.service";
 import { WebhookController } from "./webhook.controller";
+import { CapabilityController, ImportController } from "./import.controller";
+import { ImportService } from "./import.service";
 
 @Module({
-  controllers: [HealthController, RequirementController, IterationController, WebhookController],
-  providers: [RequirementService, IterationService],
+  controllers: [HealthController, RequirementController, IterationController, WebhookController, ImportController, CapabilityController],
+  providers: [RequirementService, IterationService, ImportService],
 })
 export class AppModule {}